[hu_porthu] change request logic to use IPs
If DNS caching is enabled on the operating system, the hostname is
resolved only once. This limits us to one IP address, which causes
per-process rate limiting to fail because multiple processes are
connected to the same host.
To resolve the situation, each process has a unique IP address. Using
more processes than available IP addresses will result in a warning.

This worked in a similar way when the DNS caching in the OS was
turned off, but relying on that is not reliable.