[hu_porthu] use range based API limits
API limits are expressed in a series of dates on port.hu.
The previous assumption of len(days) == number of dates available is
false. The first few days are in the past (currently 2) and others are
today and in the future (1+12). Because the currently known maximum (15)
is larger than the advertised, the log severity is reduced to info to
avoid unnecessary warnings.