[hu_porthu]: implement radio api support
The radio api as very similar to the tv one.
The provided information is very scare at the moment.

The tv channel ids are kept the same. The new radio channel ids
are in the *.radio.port.hu format.

There are some small structural differeces between the tv and the
radio api that are mentioned in the comments.

The main differences:
 -  event id format, ratio event id does not contain the channel number.
    This is provided in a spearate channelId
 -  "restrictions" is an empty list
 - different "channel" object handling